Complete refactor
This refactor seperates the game specific code from code which is running in the engine context.
In the process of this refactor, the engine initialisation code had to be completely rewritten,
alongside the engine instance, game loop, and in the future, the renderer.

Everything is a mess from this refactor and will be improved over iterations.

A full changelog can be seen below:
* Seperate engine and game related functionality. Engine components are now fetched via
  methods from the engine instance.

* Create engine class which exposes engine functionality to engine implementations

* Rewrite initialisation code to expose SDL window and renderer by shared pointers.

* Refactor assets to take an engine instance

* Refactor window and renderer to no longer be static

* Add game loop class which manages and exposes the game loop

* Refactor map class to be in game subdirectory

* Refactor main into its own subdirectory

* Refactor makefile to include all subdirectories

* Begin to start including files using the include-what-you-see principle

* Begin refactor to Google/Disguise style

* Refactor texture manager into legacy

* Add engine forward-declare header

* Probably more